% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve time series data}
\usage{
get_dataset(subject_codes, family_codes, family_nbrs, series_codes,
  subject_keywords, family_keywords, series_keywords, interval, offset,
  limit = 100, drop, head, tail, server = getOption("snzts.url"))
}
\arguments{
\item{subject_codes}{List of subject codes, e.g. \code{list("HLF", "LCI", "BOP", "CPI")}.}

\item{family_codes}{List of family codes, e.g. \code{list("SA")}.}

\item{family_nbrs}{List of family numbers.}

\item{series_codes}{List of series code, e.g. \code{list("HLFQ.SAA1AZ")}}

\item{subject_keywords}{Search terms for subject.}

\item{family_keywords}{Search terms for family.}

\item{series_keywords}{Search terms for individual series.}

\item{interval}{Interval. 1 denotes monthly, 3 denotes quarterly, and 12 denotes annual.}

\item{offset}{Offset month. For example, 3 denotes the March quarter for quartely data.}

\item{limit}{Series limit. Limits the number of results returned.}

\item{drop}{Used in conjunction with limit to access results in batches. For example, setting limit to 100 and drop to 100 would yield results 101 through 200.}

\item{head}{How many values at the head of each series to keep.}

\item{tail}{How many values at the tail of each series to keep.}

\item{server}{Server URL.}
}
\value{
A data frame containing time series data in long format.  Associated metadata
(subject title, series title, etc.) is stored in an attribute named \code{metadata}.
}
\description{
Retrieve time series data matching query parameters.
}
\examples{
# Labour force status by sex and age--female, not in labour force, total all ages, quarterly, last 12 quarters.
x <- get_dataset(subject_code = "HLF", family_code = "SA",
                 series_keywords = list("female", "not in labour force", "all ages"),
                 interval = 3, tail = 12)
attr(x, "metadata")
}
