% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{get_families}
\alias{get_families}
\title{List families}
\usage{
get_families(subject_code, family_code, family_nbr, subject_keywords,
  family_keywords, server = getOption("snzts.url"))
}
\arguments{
\item{subject_code}{Subject code, e.g. \code{"HLF"}.}

\item{family_code}{Family code, e.g. \code{"SA"}.}

\item{family_nbr}{Family number.}

\item{subject_keywords}{Search terms for subject.}

\item{family_keywords}{Search terms for family.}

\item{server}{Server URL.}
}
\description{
List available time series families, where families are subgroups below subject.
}
\examples{
get_families(subject_code = "HLF", family_code = "SA", family_nbr = 7)
get_families(subject_code = "BLD", family_keywords = "region")
}
