% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{get_info}
\alias{get_info}
\title{List series}
\usage{
get_info(subject_codes, family_codes, family_nbrs, series_codes,
  subject_keywords, family_keywords, series_keywords, interval, offset,
  limit = 100, drop, server = getOption("snzts.url"))
}
\arguments{
\item{subject_codes}{List of subject codes, e.g. \code{list("HLF", "LCI", "BOP", "CPI")}.}

\item{family_codes}{List of family codes, e.g. \code{list("SA")}.}

\item{family_nbrs}{List of family numbers.}

\item{series_codes}{List of series code, e.g. \code{list("HLFQ.SAA1AZ")}}

\item{subject_keywords}{Search terms for subject.}

\item{family_keywords}{Search terms for family.}

\item{series_keywords}{Search terms for individual series.}

\item{interval}{Interval. 1 denotes monthly, 3 denotes quarterly, and 12 denotes annual.}

\item{offset}{Offset month. For example, 3 denotes the March quarter for quartely data.}

\item{limit}{Series limit. Limits the number of results returned.}

\item{drop}{Used in conjunction with limit to access results in batches. For example, setting limit to 100 and drop to 100 would yield results 101 through 200.}

\item{server}{Server URL.}
}
\description{
List available series, including basic metadata.
}
\examples{
get_info(subject_code = "HLF", family_code = "SA",
         series_keywords = list("female", "not in labour force", "all ages"))
}
