% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{get_subjects}
\alias{get_subjects}
\title{List subjects}
\usage{
get_subjects(subject_code, subject_keywords,
  server = getOption("snzts.url"))
}
\arguments{
\item{subject_code}{Subject code, e.g. \code{"HLF"}.}

\item{subject_keywords}{Search terms for subject.}

\item{server}{Server URL.}
}
\description{
List available time series subjects.
}
\examples{
get_subjects()
get_subjects(subject_keywords = list("household", "labour force"))
}
