% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{set_server}
\alias{set_server}
\title{Set server URL}
\usage{
set_server(server)
}
\arguments{
\item{server}{Server URL.}
}
\description{
Set option \code{snzts.url} which provides the default server address.
}
