% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{test_server}
\alias{test_server}
\title{Test URL}
\usage{
test_server(server = getOption("snzts.url"))
}
\arguments{
\item{server}{Server URL.}
}
\description{
Test whether or not a particular URL is a valid web service or not.
}
\examples{
test_server("http://localhost:9000/snzts/v1")
}
