% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snzts.R
\name{get_series}
\alias{get_series}
\title{Retrieve time series by series codes}
\usage{
get_series(series_codes, server = getOption("snzts.url"))
}
\arguments{
\item{series_codes}{List of series identifiers, e.g. \code{list("HLFQ.SAA3AZ", "HLFQ.S1A3S", "HLFQ.S4A3S")}.}

\item{server}{Server URL.}
}
\value{
A data frame containing time series data in long format.  Associated metadata
(subject title, series title, etc.) is stored in an attribute named \code{metadata}.
}
\description{
Retrieve time series data for a list of series codes.
}
\examples{
x <- get_series(series_codes = c("HLFQ.SAA3AZ", "HLFQ.S1A3S", "HLFQ.S4A3S"))
attr(x, "metadata")
}
